+++ com.unity.render-pipelines.lightweight/Shaders/UnlitInput.hlsl
@@ -5,10 +5,10 @@
 
 CBUFFER_START(UnityPerMaterial)
 float4 _MainTex_ST;
-half4 _Color;
-half _Cutoff;
-half _Glossiness;
-half _Metallic;
+float4 _Color;
+float _Cutoff;
+float _Glossiness;
+float _Metallic;
 CBUFFER_END
 
 #endif
